+++ sdk/core/azure-core/src/http/winhttp/win_http_transport.cpp
@@ -346,6 +346,17 @@ void WinHttpTransport::CreateRequestHandle(std::unique_ptr<_detail::HandleManage
     {
       GetErrorAndThrow("Error while setting client cert context to ignore..");
     }
+
+    // Enforce TLS version 1.2
+    auto tlsOption = WINHTTP_FLAG_SECURE_PROTOCOL_TLS1_2;
+    if (!WinHttpSetOption(
+            handleManager->m_requestHandle,
+            WINHTTP_OPTION_SECURE_PROTOCOLS,
+            &tlsOption,
+            sizeof(tlsOption)))
+    {
+      GetErrorAndThrow("Error while enforcing TLS 1.2 for connection request.");
+    }
   }
 
   if (m_options.IgnoreUnknownServerCert)